kick off repo with some configs and a README
I was blabbing about all this on Twitter and decided to grab
my laptop and get this stuff down on repo, as they say.

The README attempts to say what this is all about so I won't repeat it here.

These configs are reproductions with cleanups from a rig I put together
last week and I haven't even tested this yet because this commit turned
out to be the best way to get this code between machines.

There are a bunch of TODOs in these files and the README has some more.

Almost all of these configs are based on examples from the source or the
Grafana + OpenTelemetry documentation.